backend/requirements.txt

Django==5.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
requests==2.31.0
setuptools<81

# Integração com APIs externas
google-generativeai==0.3.2
celery==5.3.6
redis==5.0.1
